backend/requirements.txt
Django>=6.0
djangorestframework>=3.15
django-filter>=24.2
psycopg2-binary>=2.9
python-dotenv>=1.0
